% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{busiest_routes}
\alias{busiest_routes}
\title{Find busiest airport or city pairs}
\usage{
busiest_routes(dataframe, origincol, destcol)
}
\arguments{
\item{dataframe}{data to use}

\item{origincol}{origin}

\item{destcol}{destination}
}
\value{
sorted data frame with most popular routes at the top. origin and dest become
airport1 and airport2 columns in alphabetical order.
}
\description{
This finds the busiest routes.
}
\details{
It treats both directions of the route the same, so JFK-SFO and SFO-JFK are summed together.
}
