% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load DB1B data}
\usage{
load_data(datafile, cityfile, carrierfile)
}
\arguments{
\item{datafile}{path to file with DB1B data}

\item{cityfile}{path to file with city data}

\item{carrierfile}{path to data with carrier names}
}
\value{
joined DB1B data
}
\description{
Load and join DB1B data with carrier and city names.
}
\details{
Loads DB1B dataset, and joins in origin and destination city as well as ticketing and
operating carrier names
}
