% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{market_shares}
\alias{market_shares}
\title{Compute airline market shares}
\usage{
market_shares(dataframe, carriercol, origincol)
}
\arguments{
\item{dataframe}{data to use}

\item{carriercol}{column containing carrier name/ID}

\item{origincol}{column containing origin airport ID}
}
\value{
Market shares as a data frame by airport and airline.
}
\description{
Use passenger data to compute the percentage of passengers traveling on each airline
at each airport.
}
\details{
Use by specifying a dataset, and columns for the airline and origin.
}
